# src/main/resources/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TGM Schulplaner API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/calendar/{id}": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get calendar by id",
        "operationId": "getCalendar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "put": {
        "tags": [
          "calendar"
        ],
        "summary": "Modify calendar by id",
        "operationId": "modifyCalendar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyCalendar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "calendar"
        ],
        "summary": "Delete calendar by id",
        "operationId": "deleteCalendar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/calendar/{id}/entries": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get all entries for a calendar by id, filtered by start and end",
        "operationId": "getCalendarEntries",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEntry"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/user/calendars": {
      "get": {
        "tags": [
          "user",
          "calendar"
        ],
        "summary": "Get all calendars that belong to the authenticated user",
        "operationId": "getCalendarsForUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Calendar"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/user/calendar": {
      "post": {
        "tags": [
          "user",
          "calendar"
        ],
        "summary": "Create new calendar that belongs to the group if the authenticated user has access",
        "operationId": "createCalendarForUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCalendar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/calendar/{calendarId}/entry/{id}": {
      "put": {
        "tags": [
          "calendar"
        ],
        "summary": "Modify an entry for a calendar by id",
        "operationId": "modifyCalendarEntry",
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyCalendarEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "calendar"
        ],
        "summary": "Create a new entry for a calendar by id",
        "operationId": "deleteCalendarEntry",
        "parameters": [
          {
            "name": "calendarId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group/{gid}/calendars": {
      "get": {
        "tags": [
          "group",
          "calendar"
        ],
        "summary": "Get all calendars that belong to the group if the authenticated user has access",
        "operationId": "getCalendarsForGroup",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Calendar"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/group/{gid}/calendar": {
      "post": {
        "tags": [
          "group",
          "calendar"
        ],
        "summary": "Create new calendar that belongs to the group if the authenticated user has access",
        "operationId": "createCalendarForGroup",
        "parameters": [
          {
            "name": "gid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCalendar"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
    "/api/v1/calendar/{id}/entry": {
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Create a new entry for a calendar by id",
        "operationId": "createCalendarEntry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCalendarEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEntry"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer-key": [

            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ModifyCalendar": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Calendar": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CalendarEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "calendar": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NewCalendar": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ModifyCalendarEntry": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NewCalendarEntry": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer-key": {
        "type": "http",
        "name": "bearer-key",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      }
    }
  }
}